Bug Fix: Wrong Client Socks Index
In the previous code, when the client connects the server puts
its fd into a 'client_socks list', list that holds all fds
used at the moment.

It happens that when some user disconnects, an empty slot may
occur in the middle of the list, instead of always of the end.
In order to properly handle this case, wsServer also needs to
iterate over the list and thus, finds an appropriate empty slot
to use, if any.

This commit fixes this issue and adjusts to the expected
behavior.