Feature: Initial support to Doxygen and m.css
Although wsServer is quite simple, I feel bad about receiving
contributions to the project and not even having official
documentation of the project. Therefore, this commit introduces
support for Doxygen and m.css, as a way to generate online
documentation.

In future commits support for man-pages, pkgconfig and
installation rules will also be introduced in the makefile.