Bug Fix: Handle large messages between multiples reads
To date, wsServer assumed that every received frame could be fully
read with a single call to read(), which may not be true if the
frame is larger than the MESSAGE_LENGTH.

In addition, wsServer also assumed that each read() could have
only a single frame, which is also not true, since the client can
send multiple frames in sequence even before the server processes
the previous one.

Therefore, this commit fixes these two scenarios. However, it is
important to note that wsServer does not yet support non-FIN
frames, that is, continuation frames, opcode 0. Support for this
will be added later.