Bug Fix: Allows websocket header fragmentation
The previous commit, though it adds support for continuation
frames, implementation had some issues and added rather
unnecessary complexity to the code.

Although it handled message fragmentation well, it did not
take into account a possible fragmentation of the websocket
header itself, which would lead to incorrect parsing of
subsequent frames.

This commit, therefore, fixes the above issue by treating
frame parsing as a data stream (rather than trying to read
the entire header at once) that is dynamically read by the
next_byte() routine. This small change could significantly
reduce the complexity of the code and be flexible enough to
read the frame independent of the sequence of reads required
for this.