Enhancement: Adds maximum limit for frame size
To date, wsServer did not impose a maximum limit on the frame size.
It turns out that RFC specifies that a frame can be up to 2^64 bytes,
and we know that we still don't have systems to process such a massive
amount of data, at least not popularly available.

So we need to limit the maximum amount of data that wsServer can receive
from a client, and if it exceeds, aborts the connection. Ignoring the
message is not an option, as this may constitute an attack on the server.