Feature: Fuzzing Support
This commit adds fuzzing support by implementing the ws_file routine,
which reads a set of frames from a file as if they came from the
network. This new routine allows wsServer to be 'fuzzed' by the AFL,
which generates new inputs from an existing set and checks the
stability of the program, be it crashes or timeout.