Enhancement: Implemented compatibility of send()/recv() with AFL_FUZZ
Compilation with AFL_FUZZ transforms the fd handle, which normally represents a socket, into a local file descriptor.

The send() function expects the handle of a socket as a parameter and presents an error if the handle of a local file is provided.
Therefore, when using AFL_FUZZ, the use of the write() function is maintained.