Bug Fix: Wrong Client fd For Broadcast
In #20, it was observed that the wsServer was incorrectly sending
the broadcast message to itself, instead of the remaining clients.
This is due to using the same client fd as the destination, instead
of using the fd for each connected client. Fix this by changing
'fd' to 'sock'.

Thanks @lgg2 for reporting this.