Enhancement: Non-blocking ws_socket option
In issue #24, I get asked about the possibility of the 'ws_socket'
routine not blocks the execution of the program. So far, the
alternative would be to create a separate thread and then invoke the
ws_socket from it. This commit adds a third argument to the ws_socket
function, called thread_loop, that when 0, performs the default
behavior, i.e., blocks the main thread. When != 0, the loop to accept
connections is then executed from a new thread.