Bug Fix: Fix issue #40
This commit fixes issue #40 where the Artillery tool was able to
force a race condition that occurred on the wsServer (a) and also
the premature sending of broadcast messages (b).

In detail:
a) There was a race condition during the disconnect process, where
the 'client_socks' structure was not fully secured with locks, and
wsServer tried to send messages to already disconnected clients.

b) During the broadcast of messages, the wsServer did not check if
the connected clients were already in the 'OPEN' state, that is,
if they had already performed the handshake, which made the server
send messages prematurely to the clients.

Fixed these two issues, now wsServer is able to pass Artillery tests
without problems as described in the issue.

Signed-off-by: Davidson Francis <davidsondfgl@gmail.com>