Bug Fix: Race Condition of Multiple Messages for the Same Fd
In discussion #43, it was noted that when sending multiple frames from
different threads to the same client, there could be a chance of a
race condition and the sent messages getting mixed up, which is
clearly not desirable.

This commit fixes this by serializing all 'send's related to the same
fd, thus ensuring the order of all messages.