Bug Fix: Fix issue #41
In issue #41 there was a problem getting the IP address during the
onclose() event, as the server was trying to get an IP from an
already disconnected client.

This commit fixes this by saving the IP address of the client right
during the initial stages of the connection, which makes it available
at any time, even when the client is disconnected.