Bug Fix: Multiple events for multiple ports
According to issue #85, wsServer always uses the last recorded
event even when the server listens on multiple ports with
multiple different events.

After a detailed inspection, it was possible to confirm that the
server actually saves events globally and thus only keeps one in
memory at a time. This commit fixes this by keeping events as
part of the client connection.