Bug Fix: HTTP Headers Comparison Should Ignore Case
In issue #98, it was brought to my attention that according to RFC7540
(HTTP/2), header comparisons must be case-insensitive. However, this was
not happening in `wsServer`, which caused failures in certain scenarios,
such as when used with HAProxy, which sends headers in lowercase.

This commit addresses the issue by implementing a case-insensitive
comparison when searching for the string 'Sec-WebSocket-Key'.