Bug Fix: Remove unneeded call to ws_sendframe()
The bug was introduced in PR #96, which added support for a client UID.

The issue is that by adding a "UID," it also necessitates looping through
the list of clients to find the connection corresponding to the specific
UID. This loop needs to be protected by a mutex to prevent modifications
to the client list while it's being traversed, and this part works
perfectly fine.

The problem arises when a ping broadcast is sent to all active clients.
In that case, the client list is also locked with the same mutex, causing
the code to enter a blocking state.

The fix is quite simple: switch from using `ws_sendframe()` to
`ws_sendframe_internal()`, since the latter does not rely on the client's
UID but instead uses the connection structure, which was already obtained
earlier. This way, no additional locks are needed, and the code runs
without issues.